Fix service interface namespaces and ViewModels - Major compilation error reduction
- Fixed incorrect Identity namespace references across all controllers
- Updated service interface using statements to use correct namespaces:
  - ICurrentUserService, IInteractionService, IProfileService, etc.
- Fixed ViewModels namespace references in controllers
- Created missing SettingsVM for Dashboard settings
- Fixed BackgroundJobs and Cache ViewModels namespace references
- Fixed ambiguous ChangePasswordVM reference in SecurityController
- Updated _ViewImports.cshtml with correct namespace references
- Fixed TagHelpers service interface references
- Reduced compilation errors from 215 to approximately 150+ errors
- Application layer now compiles successfully with warnings only

+++ src/CommunityCar.Application/Services/Dashboard/Maintenance/MaintenanceService.cs
@@ -141,7 +141,7 @@ public class MaintenanceService : IMaintenanceService
             EmergencyMaintenanceCount = random.Next(5, 25),
             AverageDowntimeMinutes = random.Next(60, 180),
             TotalDowntimeHours = random.Next(100, 500),
-            SuccessfulMaintenanceRate = (decimal)(0.85 + random.NextDouble() * 0.1), // 85-95%
+            SuccessfulMaintenanceRate = (double)(0.85 + random.NextDouble() * 0.1), // 85-95%
             AverageNotificationsSent = random.Next(500, 2000),
             LastMaintenanceDate = DateTime.UtcNow.AddDays(-random.Next(1, 30)),
             NextScheduledMaintenance = DateTime.UtcNow.AddDays(random.Next(7, 30)),
